Refactor - Move creation of layers to its own class.  Better map handling.
