+++ src/app/layers.spec.ts
@@ -1,0 +1,7 @@
+import { Layers } from './layers';
+
+describe('Layers', () => {
+  it('should create an instance', () => {
+    expect(new Layers()).toBeTruthy();
+  });
+});
